refactor: compose direct runner stages
